+++ src/cms/cms.js
@@ -3,8 +3,4 @@ import uploadcare from 'netlify-cms-media-library-uploadcare'
 import AboutPagePreview from './preview-templates/AboutPagePreview'
 
 CMS.registerMediaLibrary(uploadcare)
-
-
-
 CMS.registerPreviewTemplate('about', AboutPagePreview)
-
