+++ src/cms/preview-templates/AboutPagePreview.js
@@ -18,5 +18,3 @@ AboutPagePreview.propTypes = {
 }
 
 export default AboutPagePreview
-
-
