+++ src/components/all.sass
@@ -1,8 +1,6 @@
 @import "~bulma/sass/utilities/initial-variables"
 
-
 // Config vars
-
 $body-color: #333
 $black: #2b2523
 $dark-grey: #292F33
@@ -197,7 +195,6 @@ footer
   font-family: Glacial Indifference
   align-items: center
 
-
 .home-cv-item-logo
   margin-right: 1rem
   font-size: 1.5rem
@@ -259,7 +256,6 @@ footer
     background-position: center
     margin: 1rem auto 1rem
     
-
 .btn
     display: inline-block
     padding: 12px 16px 10px
@@ -343,7 +339,6 @@ footer
   font-weight: 500
   font-family: Montserrat Light
 
-
 //Menu overrides
 $menu-item-color: $white-ter
 $menu-item-hover-color: $black !important
@@ -352,7 +347,6 @@ $menu-item-active-color: $primary-color
 $menu-item-active-background-color:	$primary-color
 $menu-list-border-left: 1px solid $primary-color
 $menu-label-color: $white-ter
-
 
 .menu-label
   font-size: 1em !important
