+++ src/components/home/Avatar.js
@@ -35,5 +35,4 @@ const Avatar = () => {
     </div>
   )}
 
-  
 export default Avatar
