+++ src/pages/index.js
@@ -1,11 +1,10 @@
 import React from 'react'
 import Layout from '../components/Layout'
 import getJokes from '../dad_jokes/dadJokes'
-import Bio from '../components/Bio'
-import ContactMe from '../components/ContactMe'
-
-import EducationList  from '../components/EducationList'
-import Avatar from "../components/Avatar"
+import Bio from '../components/home/Bio'
+import ContactMe from '../components/home/ContactMe'
+import EducationList  from '../components/home/EducationList'
+import Avatar from "../components/home/Avatar"
 
 export default class Index extends React.Component {
   constructor(props) {
