+++ src/pages/shinies/index.js
@@ -1,5 +1,4 @@
 import React from 'react'
-
 import Layout from '../../components/Layout'
 import ShinyRoll from '../../components/ShinyRoll'
