+++ src/styles/global.css
@@ -36,7 +36,3 @@
     font-weight: 300;
     font-style: italic;
 }
-
-/* h1, h3, body {
-    font-family: Glacial Indifference;
-} */
